+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessExtensionImpl.java
@@ -17,7 +17,7 @@ package com.diffplug.gradle.spotless;
 
 import org.gradle.api.Action;
 import org.gradle.api.Project;
-import org.gradle.api.plugins.BasePlugin;
+import org.gradle.api.UnknownTaskException;
 import org.gradle.api.plugins.JavaBasePlugin;
 import org.gradle.api.tasks.TaskContainer;
 import org.gradle.api.tasks.TaskProvider;
@@ -43,8 +43,12 @@ public class SpotlessExtensionImpl extends SpotlessExtension {
 
 		project.afterEvaluate(unused -> {
 			if (enforceCheck) {
-				project.getTasks().named(JavaBasePlugin.CHECK_TASK_NAME)
-						.configure(task -> task.dependsOn(rootCheckTask));
+				try {
+					project.getTasks().named(JavaBasePlugin.CHECK_TASK_NAME)
+							.configure(task -> task.dependsOn(rootCheckTask));
+				} catch (UnknownTaskException e) {
+					// no action needed, it's okay if there's no `check` task
+				}
 			}
 		});
 	}
@@ -78,17 +82,14 @@ public class SpotlessExtensionImpl extends SpotlessExtension {
 	protected void createFormatTasks(String name, FormatExtension formatExtension) {
 		boolean isIdeHook = project.hasProperty(IdeHook.PROPERTY);
 		TaskContainer tasks = project.getTasks();
-		TaskProvider<?> cleanTask = tasks.named(BasePlugin.CLEAN_TASK_NAME);
 
 		// create the SpotlessTask
 		String taskName = EXTENSION + SpotlessPlugin.capitalize(name);
 		TaskProvider<SpotlessTaskImpl> spotlessTask = tasks.register(taskName, SpotlessTaskImpl.class, task -> {
 			task.init(getRegisterDependenciesTask().getTaskService());
 			task.setEnabled(!isIdeHook);
-			// clean removes the SpotlessCache, so we have to run after clean
-			task.mustRunAfter(cleanTask);
 		});
-
+		SpotlessPlugin.taskMustRunAfterClean(project, spotlessTask);
 		project.afterEvaluate(unused -> {
 			spotlessTask.configure(task -> {
 				// now that the task is being configured, we execute our actions
@@ -129,8 +130,8 @@ public class SpotlessExtensionImpl extends SpotlessExtension {
 		// create the diagnose task
 		TaskProvider<SpotlessDiagnoseTask> diagnoseTask = tasks.register(taskName + DIAGNOSE, SpotlessDiagnoseTask.class, task -> {
 			task.source = spotlessTask.get();
-			task.mustRunAfter(cleanTask);
 		});
+		SpotlessPlugin.taskMustRunAfterClean(project, diagnoseTask);
 		rootDiagnoseTask.configure(task -> task.dependsOn(diagnoseTask));
 	}
 }