+++ plugin-maven/CHANGES.md
@@ -4,6 +4,11 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 
 ## [Unreleased]
 
+## [2.17.6] - 2021-12-05
+### Changed
+* Bumped default ktlint from `0.43.0` to `0.43.2`.
+
+## [2.17.5] - 2021-12-01
 ### Changed
 * Bump jgit version ([#992](https://github.com/diffplug/spotless/pull/992)).
   * jgit `5.10.0.202012080955-r` -> `5.13.0.202109080827-r`